fix：trash file was deleted when restoring conflict and select skip
8346 回收站还原文件，关闭重复文件提示框之后，回收站的文件再次还原失败  来源Bug-1638
当源文件处于桌面以外的其他数据盘时，在还原对话框中点击跳过，回收站中的文件会被删除

+++ fileoperations/filejob.cpp
@@ -586,7 +586,7 @@ bool FileJob::doTrashRestore(const QString &srcFilePath, const QString &tarFileP
                 ok = QFile::rename(_tarFilePath, tarFilePath);
             }
         }else if (srcInfo.isFile() || srcInfo.isSymLink()){
-            if (copyFile(srcFilePath, tarDir, true, &_tarFilePath)) {
+            if (copyFile(srcFilePath, tarDir, true, &_tarFilePath) && !getIsSkip()) {
                 deleteFile(srcFilePath);
                 ok = QFile::rename(_tarFilePath, tarFilePath);
             }
